refactored OSUtils
removed os utils methods moved to functions.java
