+++ Address Book/src/com/COP2805C/AddressBook/UserInterface/MainWindow.java
@@ -14,10 +14,7 @@ import javafx.scene.layout.FlowPane;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 
-import javax.swing.plaf.basic.BasicToolTipUI;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 
 /**
  * Created by abirfaisal on 6/8/15.
@@ -141,12 +138,12 @@ public class MainWindow {
             }
 
             if (newValue.contains("Last")){
-                Collections.sort(Main.getContactInformationArrayList(), new LastNameComparator());
+                Collections.sort(Main.getContactInformationArrayList(), new Functions.LastNameComparator());
                 Functions.refreshListView();
             }
 
             if (newValue.contains("Phone #")){
-                Collections.sort(Main.getContactInformationArrayList(), new PhoneNumberComparator());
+                Collections.sort(Main.getContactInformationArrayList(), new Functions.PhoneNumberComparator());
                 Functions.refreshListView();
             }
         });
@@ -261,31 +258,6 @@ public class MainWindow {
         return menuButton;
     }
 
-    //Sorting comparator by LastName
-    static class LastNameComparator implements Comparator<ContactInformation>{
-        @Override
-        public int compare(ContactInformation o1, ContactInformation o2) {
-            return o1.getLastName().compareToIgnoreCase(o2.getLastName());
-        }
-    }
-    //Sorting comparator by PhoneNumber
-    static class PhoneNumberComparator implements Comparator<ContactInformation>{
-        @Override
-        public int compare(ContactInformation o1, ContactInformation o2) {
-            Long contact1, contact2;
-                if(o1.getPhoneNumbers().size()==0){
-                    contact1 = 0L;
-                }else{
-                    contact1 = Long.parseLong(o1.getPhoneNumbers().get(0).replace("-",""));
-                }
-                if(o2.getPhoneNumbers().size()==0){
-                    contact2 = 0L;
-                }else{
-                    contact2 = Long.parseLong(o2.getPhoneNumbers().get(0).replace("-",""));
-                }
-                return Long.compare(contact1,contact2);
-        }
-    }
 }
